Convert between Temperature in Degrees C and Proxy Units

A wrapper function for accessing proxy - temperature conversion functions

Usage

ProxyConversion(
  temperature = NULL,
  proxy.value = NULL,
  calibration.type = "identity",
  slp.int.means = NULL,
  slp.int.vcov = NULL,
  calibration = switch(calibration.type, identity = NA, Uk37 = "Mueller global", MgCa =
    "Ten planktonic species_350-500"),
  point.or.sample = c("point", "sample"),
  n = 1
)

Arguments

temperature

Temperature in degrees C

proxy.value

Temperature in proxy units

calibration.type

Type of proxy, e.g. Uk37 or MgCa

slp.int.means

Optional user supplied vector of values for the slope and intercept of the calibration function. Overides the defaults.

slp.int.vcov

Optional user supplied variance covariance matrix calibration parameters. Overides the defaults.

calibration

The name of a specific calibration for which calibration parameters are provided by sedproxy. Currently applies only to calibration.type MgCa.

point.or.sample

Use the "best estimate" calibration parameters, or parameters sampled from the fitted calibration model

n

the number of replicate conversions to make in the case of sampled calibration parameters

Value

a vector of temperatures or proxy values

Details

Valid entries for calibration are: "Ten planktonic species_350-500", "G. aequilateralis_350-500", "G. aequilateralis_500-1000", "G. conglobatus_350-500", "G. hirsuta_350-500", "G. inflata_350-500", "G. ruber pink_250-350", "G. ruber pink_350-500", "G. ruber white_250-350", "G. ruber white_350-500", "G. sacculifer with sac_350-500", "G. sacculifer without sac_350-500", "G. truncatulinoides_350-500", "G. truncatulinoides_500-1000", "N. dutertrei_350-500", "O. univesa_350-500", "P. obliquiloculata_350-500"

Examples

# From temperature to UK'37
## With fixed calibration
ProxyConversion(temperature = c(10, 20), point.or.sample = "point",
                calibration.type = "Uk37")
#> [1] 0.374 0.704

## With random calibration, 5 replicates
ProxyConversion(temperature = c(1, 2), n = 5, point.or.sample = "sample",
                calibration.type = "Uk37")
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.06692328 0.07517807 0.05834467 0.0837350 0.08155695
#> [2,] 0.10050146 0.10813490 0.09234955 0.1164315 0.11437804

## Back-transformation with same calibration
ProxyConversion(
  proxy.value = as.vector(
    ProxyConversion(
      temperature = c(21, 22),
      calibration.type = "Uk37",
      point.or.sample = "point"
    )
  ),
  point.or.sample = "point",
  calibration.type = "Uk37"
)
#> [1] 21 22

## Back-transformation with random calibration
ProxyConversion(
  proxy.value = as.vector(
    ProxyConversion(
      temperature = c(21, 22),
      calibration.type = "Uk37",
     point.or.sample = "point"
    )
  )
  ,
  n = 5,
  point.or.sample = "sample",
  calibration.type = "Uk37"
)
#>          [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,] 20.92230 20.91001 21.18363 21.06548 20.86180
#> [2,] 21.90798 21.91211 22.19361 22.07434 21.85299

## Incompatible arguments
if (FALSE) { # \dontrun{
ProxyConversion(temperature = 1, proxy.value = 1)
} # }