Description of proxy stages

A description of the proxy stages in the output of ClimToProxyClim and default labels, colours and order for plotting

Usage

stages.key

Format

A data frame with 13 rows and 6 variables:

stage

proxy stage	label

label for proxy stage	description

description of proxy stage	plot.order

default plotting order of stages	plotting.colour

default colour for plotting	plotting.alpha